context for UXCam
