+++ client/app/(tabs)/settings/index.tsx
@@ -5,8 +5,6 @@ import SettingsItem from '../../../components/SettingsItem';
 import RecordingButton from '../../../components/RecordingButton';
 
 export default function SettingsScreen() {
-  const [isRecording, setIsRecording] = useState(false);
-
   return (
     <>
       <Stack.Screen
@@ -23,7 +21,7 @@ export default function SettingsScreen() {
           <SettingsItem title="Language" route="/settings/language" />
           <SettingsItem title="Text Size" route="/settings/textsize" />
           <SettingsItem title="Color Blindness" route="/settings/colorblind" />
-          <RecordingButton isRecording={isRecording} setIsRecording={setIsRecording} />
+          <RecordingButton />
         </View>
       </View>
     </>
